+++ src/textual/dom.py
@@ -12,6 +12,7 @@ from functools import lru_cache
 from inspect import getfile
 from typing import (
     TYPE_CHECKING,
+    Callable,
     ClassVar,
     Iterable,
     Sequence,
@@ -49,6 +50,7 @@ if TYPE_CHECKING:
     from rich.console import RenderableType
     from .app import App
     from .css.query import DOMQuery, QueryType
+    from .message import Message
     from .screen import Screen
     from .widget import Widget
     from .worker import Worker, WorkType, ResultType
@@ -146,6 +148,8 @@ class DOMNode(MessagePump):
     _merged_bindings: ClassVar[_Bindings | None] = None
 
     _reactives: ClassVar[dict[str, Reactive]]
+
+    _decorated_handlers: dict[type[Message], list[tuple[Callable, str | None]]]
 
     def __init__(
         self,